imporvements

+++ Code/TranslationUtil.py
@@ -16,7 +16,11 @@ class Translator_Util:
     def __init__(self):
         self.translator = Translator()
         self.helper = Helper()
-        self.device = Config.get_device()        
+        self.device = Config.get_device()
+        if self.device == 'DMM':
+            self.masters_path = Path(Paths.GAME_MASTERS_DMM)
+        elif self.device == 'Android':
+            self.masters_path = Path(Paths.GAME_MASTERS_Android_Local)       
         self.source_path = Path(self.device) / Path(Paths.SOURCE_DIR)        
         self.source_path.mkdir(parents=True, exist_ok=True)
         self.updated_files_path = Path(self.device) / Path(Paths.UPDATED_FILES_DIR)        
@@ -25,6 +29,8 @@ class Translator_Util:
         self.new_entries_path.mkdir(parents=True, exist_ok=True)
         self.translated_files_path = Path(self.device) / Path(Paths.TRANSLATED_FILES_DIR)        
         self.translated_files_path.mkdir(parents=True, exist_ok=True)
+        self.masters_backup_path = Path(self.device) / Path(Paths.MASTERS_BACKUP)        
+        self.masters_backup_path.mkdir(parents=True, exist_ok=True)
 
     def __translate_file(self, filename:str, path:str):
         print(f"       ├─ 🔁 Translating file {filename}.")
@@ -239,19 +245,27 @@ class Translator_Util:
         Config.set_updated_files(updated_files)
 
         # Delete backups before generating new files
-        source_dir = Path(Paths.MASTERS_BACKUP)
+        backup_dir = Path(self.masters_backup_path)
+        for file in backup_dir.iterdir():
+            if file.is_file():
+                file.unlink()
+
+        source_dir = Path(self.new_entries_path)
         for file in source_dir.iterdir():
             if file.is_file():
                 file.unlink()
 
-        source_dir = Path(Paths.NEW_ENTRIES_DIR)
-        for file in source_dir.iterdir():
-            if file.is_file():
-                file.unlink()
+        # For android, pull updated files
+        if self.device == 'Android':
+            source_dir = Path(self.masters_path)
+            for file in source_dir.iterdir():
+                if file.is_file():
+                    file.unlink()
+            self.helper.pull_updated_files_from_mobile(timestamp)
 
         # 🔁 Walk through all files in the masters folder
-        for filename in os.listdir(Paths.GAME_MASTERS):
-            file_path = os.path.join(Paths.GAME_MASTERS, filename)
+        for filename in os.listdir(self.masters_path):
+            file_path = os.path.join(self.masters_path, filename)
 
             # Skip subfolders
             if not os.path.isfile(file_path):
